BSR : add menu tests without hoc withRouter
